// src/entities/Article/model/consts/consts.ts
export enum ArticleSortField {
    VIEWS = 'views',
    TITLE = 'title',
    CREATED = 'createdAt'
}

export enum ArticleBlockType {
    CODE = 'CODE',
    IMAGE = 'IMAGE',
    TEXT = 'TEXT'
}

export enum ArticleType {
    ALL = 'ALL',
    DRAMMA = 'DRAMMA',
    ROMANCE = 'ROMANCE',
    SCHOOL = 'SCHOOL',
    COMEDY = 'COMEDY',
    THRILLER = 'THRILLER',
    ACTION = 'ACTION',
    ADVENTURE = 'ADVENTURE',
}

export enum ArticleView {
    BIG ='BIG',
    SMALL = 'SMALL'
}